Big refactor
Adds support for running benchmark for a fixed amount of time.  If an object
(Pod/Job) finishes before that time, a new instance is started.

To support this, workloads should include a new annotation "multipleInstances",
which is a comma separated list of the file names contained in the workload which
should be restarted after they finish.  This list is also used now to know which
objects should be duplicated when a createObjSpec has a count > 1.

This is also used to automatically determine if `syncStart` should be enabled: if
count > 1 and the objects to be duplicated have init containers, then it's assumed
that their start should be synchronized.

TODO: Currently the runtime starts counting from the very beginning (i.e. when the
benchmark object is created).  It might be desirable to start counting after the
initialization phase.