Changes scaling, snapshotting, output collection.
Scaling: Done using custom scripts now.  I/O workload should include a
scaleUp.sh and scaleDown.sh, Operator runs those in a Pod it starts to
scale a workload up/down.  The Pod uses a service account that's allowed
to use kubectl to patch items, so the use case now is when you need to
query a custom object to get the current replica count then patch the
object to set the new replica count.
TODO: Add back in the ability to scale standard k8s objects like Deployments
without having to provide a script.

Instead of snapshotting a given volume, snapshots every volume that matches the
given selector.
TODO: Need to change this to lookup the volume associated with the given workload
action and only snapshot that.
TODO: Need to allow different policies for when more than one volume is found -
snapshot all, newest, random, etc.

Instead of parsing parser-container's stdout into a map now we just throw the last
line string into the results that get sent to output.  Much better, now if the
parser errored out we send the error to output rather than failing because it's not
valid json.  Also allows more flexibility in the output.