Adds output collector container.
Runs a simple http server that listens on port 8888 and outputs a json
object for each request it receives.  The server gets deployed in the
cnsbench-system namespace and is exposed via a service - you can send
output to it via http://cnsbench-output-collector.cnsbench-system.svc.cluster.local:8888.
By default, the benchmark's metadata info is sent there after a benchmark
finishes.